refactor(precompiles): move grant_role_unchecked off the Stablecoin trait
Addresses PR review feedback (#3960, r3581396116): grant_role_unchecked was a
default method on the public Stablecoin trait, which (1) can't be frozen
per-version — it's shared across versions — and (2) is reachable on
&dyn Stablecoin, widening the unauthorized-role-grant surface.

Make it an inherent pub(crate) method on StablecoinV1 (symmetric with the
private revoke_role_unchecked), so it stays frozen with V1 and off the trait
object. V1's grant_role calls it directly; the factory bootstrap in dispatch
calls the concrete StablecoinV1 (already pinned to V1).
